fix: harden decoder and calibration against edge cases and reduce duplication
- Make Calibration arrays read-only to enforce frozen dataclass contract
- Validate frac_us < 1M in timestamp parsing instead of silent truncation
- Move _decode_packet_spherical import out of public namespace
- Change SPHERICAL_DTYPE channel from uint8 to uint16 for consistency with POINT_DTYPE ring
- Extract shared block-parsing logic into _extract_blocks to eliminate decode duplication
- Try CRLF line endings before LF when validating calibration CSV checksums
- Handle big-endian pcap files in integration test reader
- Add tests/conftest.py for explicit _packet_builder sys.path setup

+++ src/lidar2numpy/structs.py
@@ -83,7 +83,7 @@ POINT_DTYPE: np.dtype[np.void] = np.dtype(
 # channel is 1-based (1–128) to match the ring field in POINT_DTYPE.
 SPHERICAL_DTYPE: np.dtype[np.void] = np.dtype(
     [
-        ("channel", np.uint8),
+        ("channel", np.uint16),
         ("azimuth_deg", np.float32),  # calibration-corrected, degrees
         ("distance_m", np.float32),  # metres
         ("intensity", np.float32),
